add proposal votes store
